Adição do componente inicial da argumentação e melhoria no carregamento dos RPs
