+++ src/utils/argumentacao.js
@@ -1,0 +1,24 @@
+const argumentacao = {
+  'Discussões sobre o tema e o escopo': [
+      {
+          id: 'RPSalehV1',
+          img: RPSalehV1,
+          autor: members.Saleh,
+          versao: 1,
+          descricao: 'Rich Picture que mostra o fluxo geral do funcionamento do Habitica.',
+          data: '10/10/2019',
+          fontes: [{nome: 'Google', link:'https://www.google.com/'}]
+      },
+      {
+          id: 'RPRogerV1',
+          img: RPRogerV1,
+          autor: members.Rogerio,
+          versao: 1,
+          descricao: 'Rich Picture que mostra o fluxo geral do funcionamento do Habitica.',
+          data: '10/10/2019',
+          fontes: [{nome: 'Google', link:'https://www.google.com/'}]
+      },
+  ],
+}
+
+export default argumentacao
